PUBDEV-8996 - Separate each test into its own class and isolate the failures as much as possible - in previous implementation - the failure in one tests could have effect on failure of other. This separation fixed the problem
