+++ h2o-extensions/xgboost/src/test/java/hex/tree/xgboost/remote/SteamExecutorStarterStartTest.java
@@ -1,0 +1,78 @@
+package hex.tree.xgboost.remote;
+
+import hex.tree.xgboost.XGBoost;
+import hex.tree.xgboost.XGBoostModel;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import water.H2O;
+import water.Job;
+import water.Scope;
+import water.fvec.Frame;
+import water.runner.CloudSize;
+import water.runner.H2ORunner;
+import water.test.WebsocketClient;
+
+import java.util.Map;
+import java.util.Optional;
+
+import static hex.tree.xgboost.remote.SteamExecutorStarterTestHelper.*;
+import static org.junit.Assert.*;
+import static water.TestUtil.parseTestFile;
+
+@RunWith(H2ORunner.class)
+@CloudSize(1)
+public class SteamExecutorStarterStartTest {
+
+    static boolean orig_external_xgboost;
+
+    @BeforeClass
+    public static void configureRemoteXGBoost() {
+        orig_external_xgboost = H2O.ARGS.use_external_xgboost;
+        H2O.ARGS.use_external_xgboost = true;
+    }
+
+    @AfterClass
+    public static void revertRemoteXGBoost() {
+        H2O.ARGS.use_external_xgboost = orig_external_xgboost;
+    }
+
+    @Test
+    public void testSteamClusterStart() throws Exception {
+        Scope.enter();
+        final WebsocketClient steam = new WebsocketClient();
+        try {
+            Frame train = Scope.track(parseTestFile("./smalldata/prostate/prostate.csv"));
+
+            XGBoostModel.XGBoostParameters params = new XGBoostModel.XGBoostParameters();
+            params._train = train._key;
+            params._response_column = "AGE";
+            params._ignored_columns = new String[]{"ID"};
+
+            // queue two model in parallel
+            Job<XGBoostModel> model1 = new XGBoost(params).trainModel();
+            Job<XGBoostModel> model2 = new XGBoost(params).trainModel();
+
+            // first request will request external cluster start
+            Map<String, String> startReq = steam.waitToReceiveMessage("start request");
+            assertNotNull(startReq.get("_id"));
+            assertEquals("startXGBoostCluster", startReq.get("_type"));
+            steam.sendMessage(makeStartingResponse(startReq));
+            steam.sendMessage(makeStartedResponse(startReq));
+            Scope.track_generic(model1.get());
+
+            // second model should not go to steam
+            Optional<Map<String, String>> startRequest2 = steam.waitToReceiveMessage("none", 1_000, false);
+            assertFalse("Unexpected message to steam", startRequest2.isPresent());
+            Scope.track_generic(model2.get());
+
+            // steam requests cluster stop
+            steam.sendMessage(makeStopReq("01_stop_req"));
+            expectAndCheckStopResponse(steam, "01_stop_req", true);
+        } finally {
+            Scope.exit();
+            steam.close();
+        }
+    }
+}
