+++ h2o-extensions/xgboost/src/test/java/hex/tree/xgboost/remote/SteamExecutorStarterTerminateNoClusterTest.java
@@ -1,0 +1,88 @@
+package hex.tree.xgboost.remote;
+
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import water.H2O;
+import water.Scope;
+import water.runner.CloudSize;
+import water.runner.H2ORunner;
+import water.test.WebsocketClient;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static hex.steam.SteamMessenger.ID;
+import static hex.steam.SteamMessenger.TYPE;
+import static org.junit.Assert.assertEquals;
+
+@RunWith(H2ORunner.class)
+@CloudSize(1)
+public class SteamExecutorStarterTerminateNoClusterTest {
+
+    static boolean orig_external_xgboost;
+
+    @BeforeClass
+    public static void configureRemoteXGBoost() {
+        orig_external_xgboost = H2O.ARGS.use_external_xgboost;
+        H2O.ARGS.use_external_xgboost = true;
+    }
+
+    @AfterClass
+    public static void revertRemoteXGBoost() {
+        H2O.ARGS.use_external_xgboost = orig_external_xgboost;
+    }
+
+    protected Map<String, String> makeStartedResponse(Map<String, String> startReq) {
+        return makeStartedResponse(startReq, "started", null);
+    }
+
+    protected Map<String, String> makeStartedResponse(Map<String, String> startReq, String status, String reason) {
+        Map<String, String> startResp = new HashMap<>();
+        startResp.put(ID, startReq.get(ID) + "_response");
+        startResp.put(TYPE, "xgboostClusterStartNotification");
+        startResp.put("uri", H2O.CLOUD.leader().getIpPortString());
+        startResp.put("status", status);
+        if (reason != null) {
+            startResp.put("reason", reason);
+        }
+        return startResp;
+    }
+
+    protected Map<String, String> makeStartingResponse(Map<String, String> startReq) {
+        Map<String, String> startResp = new HashMap<>();
+        startResp.put(ID, startReq.get(ID) + "_response");
+        startResp.put(TYPE, "xgboostClusterStartNotification");
+        startResp.put("status", "starting");
+        return startResp;
+    }
+
+    protected Map<String, String> makeStopReq(String id) {
+        Map<String, String> stopReq = new HashMap<>();
+        stopReq.put(ID, id);
+        stopReq.put(TYPE, "stopXGBoostClusterNotification");
+        return stopReq;
+    }
+
+    protected void expectAndCheckStopResponse(WebsocketClient steam, String id, boolean allowed) {
+        Map<String, String> stop2Resp = steam.waitToReceiveMessage("response to " + id);
+        assertEquals(id + "_response", stop2Resp.get("_id"));
+        assertEquals("stopXGBoostClusterConfirmation", stop2Resp.get("_type"));
+        assertEquals(Boolean.toString(allowed), stop2Resp.get("allowed"));
+    }    
+
+    @Test
+    public void testSteamClusterTerminateNoCluster() throws Exception {
+        Scope.enter();
+        final WebsocketClient steam = new WebsocketClient();
+        try {
+            // steam requests cluster stop
+            steam.sendMessage(makeStopReq("04_stop_req"));
+            expectAndCheckStopResponse(steam, "04_stop_req", true);
+        } finally {
+            Scope.exit();
+            steam.close();
+        }
+    }
+}
